# apps/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Triathlon API",
    "version": "0.2.0",
    "description": "One public HTTP interface for all Triathlon clients. Error responses share the Problem Details envelope with a stable domain code."
  },
  "components": {
    "securitySchemes": {
      "bearerKey": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/projects/{projectId}/tickets": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "descriptionMd": {
                    "type": "string"
                  },
                  "points": {
                    "minimum": 0,
                    "type": "integer"
                  },
                  "priority": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "low"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "medium"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "high"
                        ]
                      }
                    ]
                  },
                  "labels": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "assigneeId": {
                    "type": "string"
                  },
                  "parentId": {
                    "type": "string"
                  },
                  "blockerIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "columnId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "projectId",
                    "number",
                    "title",
                    "points",
                    "priority",
                    "labels",
                    "assigneeId",
                    "columnId",
                    "position",
                    "revision",
                    "resourceVersion",
                    "createdAt",
                    "deletedAt",
                    "category",
                    "descriptionMd",
                    "parentId",
                    "createdBy"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "projectId": {
                      "type": "string"
                    },
                    "number": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "points": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "priority": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "low"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "medium"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "high"
                          ]
                        }
                      ]
                    },
                    "labels": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "assigneeId": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "columnId": {
                      "type": "string"
                    },
                    "position": {
                      "type": "integer"
                    },
                    "revision": {
                      "type": "integer"
                    },
                    "resourceVersion": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "deletedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "category": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "not_started"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "started"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "done"
                          ]
                        }
                      ]
                    },
                    "descriptionMd": {
                      "type": "string"
                    },
                    "parentId": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdBy": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "open"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "closed"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "deleted"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "visibility",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "columnId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "assigneeId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "label",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "projectId",
                          "number",
                          "title",
                          "points",
                          "priority",
                          "labels",
                          "assigneeId",
                          "columnId",
                          "position",
                          "revision",
                          "resourceVersion",
                          "createdAt",
                          "deletedAt",
                          "category"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "projectId": {
                            "type": "string"
                          },
                          "number": {
                            "type": "integer"
                          },
                          "title": {
                            "type": "string"
                          },
                          "points": {
                            "anyOf": [
                              {
                                "type": "integer"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "priority": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "low"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "medium"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "high"
                                ]
                              }
                            ]
                          },
                          "labels": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "assigneeId": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "columnId": {
                            "type": "string"
                          },
                          "position": {
                            "type": "integer"
                          },
                          "revision": {
                            "type": "integer"
                          },
                          "resourceVersion": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "deletedAt": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "category": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "not_started"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "started"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "done"
                                ]
                              }
                            ]
                          }
                        }
                      }
                    },
                    "nextCursor": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/tickets/{ref}/move": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "columnId",
                  "position",
                  "expectedResourceVersion"
                ],
                "properties": {
                  "columnId": {
                    "type": "string"
                  },
                  "position": {
                    "minimum": 0,
                    "type": "integer"
                  },
                  "expectedResourceVersion": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "ref",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "projectId",
                    "number",
                    "title",
                    "points",
                    "priority",
                    "labels",
                    "assigneeId",
                    "columnId",
                    "position",
                    "revision",
                    "resourceVersion",
                    "createdAt",
                    "deletedAt",
                    "category",
                    "descriptionMd",
                    "parentId",
                    "createdBy"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "projectId": {
                      "type": "string"
                    },
                    "number": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "points": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "priority": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "low"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "medium"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "high"
                          ]
                        }
                      ]
                    },
                    "labels": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "assigneeId": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "columnId": {
                      "type": "string"
                    },
                    "position": {
                      "type": "integer"
                    },
                    "revision": {
                      "type": "integer"
                    },
                    "resourceVersion": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "deletedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "category": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "not_started"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "started"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "done"
                          ]
                        }
                      ]
                    },
                    "descriptionMd": {
                      "type": "string"
                    },
                    "parentId": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdBy": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/tickets/{ref}/comments": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "body"
                ],
                "properties": {
                  "body": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "ref",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "ticketId",
                    "authorId",
                    "body",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "ticketId": {
                      "type": "string"
                    },
                    "authorId": {
                      "type": "string"
                    },
                    "body": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/tickets/{ref}/purge": {
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "ref",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:1"
    }
  ]
}